inc(workforms): split form_questions_editor into 3 dedicated MCP tools
Split the overloaded form_questions_editor tool (which used an action
parameter to dispatch create/update/delete) into three focused tools:
- create_form_question: creates questions with type-specific schemas
- update_form_question: patch-based updates with all fields optional
- delete_form_question: simple deletion by formToken + questionId

Each tool has precise Zod schemas, specific descriptions, and appropriate
annotations (e.g. destructiveHint: true for delete). This improves LLM
tool selection accuracy by eliminating conditional required fields.

Monday: https://monday.monday.com/boards/3709356818/pulses/11453478027
