+++ packages/agent-toolkit/src/core/tools/platform-api-tools/index.ts
@@ -4,7 +4,9 @@ import { ChangeItemColumnValuesTool } from './change-item-column-values-tool';
 import { CreateBoardTool } from './create-board-tool';
 import { CreateViewTool } from './create-view-tool/create-view-tool';
 import { CreateFormTool } from './workforms-tools/create-form-tool';
-import { FormQuestionsEditorTool } from './workforms-tools/form-questions-editor-tool';
+import { CreateFormQuestionTool } from './workforms-tools/create-form-question-tool';
+import { UpdateFormQuestionTool } from './workforms-tools/update-form-question-tool';
+import { DeleteFormQuestionTool } from './workforms-tools/delete-form-question-tool';
 import { UpdateFormTool } from './workforms-tools/update-form-tool';
 import { GetFormTool } from './workforms-tools/get-form-tool';
 import { CreateColumnTool } from './create-column-tool';
@@ -68,7 +70,9 @@ export const allGraphqlApiTools: BaseMondayApiToolConstructor[] = [
   CreateFormTool,
   UpdateFormTool,
   GetFormTool,
-  FormQuestionsEditorTool,
+  CreateFormQuestionTool,
+  UpdateFormQuestionTool,
+  DeleteFormQuestionTool,
   CreateColumnTool,
   CreateGroupTool,
   DeleteColumnTool,
@@ -110,7 +114,9 @@ export * from './create-board-tool';
 export * from './workforms-tools/create-form-tool';
 export * from './workforms-tools/update-form-tool';
 export * from './workforms-tools/get-form-tool';
-export * from './workforms-tools/form-questions-editor-tool';
+export * from './workforms-tools/create-form-question-tool';
+export * from './workforms-tools/update-form-question-tool';
+export * from './workforms-tools/delete-form-question-tool';
 export * from './create-column-tool';
 export * from './create-group/create-group-tool';
 export * from './create-custom-activity-tool';
