+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workforms-tools/delete-form-question-tool/index.ts
@@ -1,0 +1,42 @@
+import { ToolInputType, ToolOutputType, ToolType } from '../../../../tool';
+import { BaseMondayApiTool, createMondayApiAnnotations } from '../../base-monday-api-tool';
+import { deleteFormQuestionToolSchema } from './schema';
+import {
+  DeleteFormQuestionMutation,
+  DeleteFormQuestionMutationVariables,
+} from '../../../../../monday-graphql/generated/graphql/graphql';
+import { deleteFormQuestion } from '../workforms.graphql';
+
+export class DeleteFormQuestionTool extends BaseMondayApiTool<typeof deleteFormQuestionToolSchema, never> {
+  name = 'delete_form_question';
+  type = ToolType.WRITE;
+  annotations = createMondayApiAnnotations({
+    title: 'Delete Form Question',
+    readOnlyHint: false,
+    destructiveHint: true,
+    idempotentHint: false,
+  });
+
+  getDescription(): string {
+    return 'Delete a question from a monday.com form. This permanently removes the question and its associated data. The question is identified by its questionId, which can be obtained from the get_form tool.';
+  }
+
+  getInputSchema(): typeof deleteFormQuestionToolSchema {
+    return deleteFormQuestionToolSchema;
+  }
+
+  protected async executeInternal(
+    input: ToolInputType<typeof deleteFormQuestionToolSchema>,
+  ): Promise<ToolOutputType<never>> {
+    const variables: DeleteFormQuestionMutationVariables = {
+      formToken: input.formToken,
+      questionId: input.questionId,
+    };
+
+    await this.mondayApi.request<DeleteFormQuestionMutation>(deleteFormQuestion, variables);
+
+    return {
+      content: { message: 'Question deleted', question_id: input.questionId, action_name: 'delete' },
+    };
+  }
+}
