+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workforms-tools/delete-form-question-tool/schema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+import { GraphQLDescriptions } from '../workforms.consts';
+
+export const deleteFormQuestionToolSchema = {
+  formToken: z.string().describe(GraphQLDescriptions.commonArgs.formToken),
+  questionId: z.string().describe(GraphQLDescriptions.commonArgs.questionId),
+};
