+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workforms-tools/workforms.types.ts
@@ -457,9 +457,3 @@ export interface Form {
   };
   tags: Tag[];
 }
-
-export enum FormQuestionActions {
-  Delete = 'delete',
-  Update = 'update',
-  Create = 'create',
-}
